requirements.txt
ccxt==4.1.13
python-dotenv==1.0.0
rich==13.7.0
loguru==0.7.2
typer==0.9.0
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2 